fix: initialize 훅 불필요한 로직 최적화
