+++ src/entities/post/api/index.ts
@@ -12,8 +12,9 @@ export const useFetchPostsQuery = (params: {
   const { searchQuery, tag, limit, skip } = params;
 
   return useQuery<PostResponse>({
-    queryKey: ["posts"],
+    queryKey: ["posts", params],
     queryFn: async () => {
+      console.log("post 요청", params);
       if (searchQuery) {
         return axios.get(`/api/posts/search?q=${searchQuery}`).then((res) => res.data);
       }
