+++ src/pages/PostsManagerPage.tsx
@@ -2,12 +2,12 @@ import { Card, CardContent } from "@/shared/card/ui";
 import PostTable from "../widgets/post/PostTable.tsx";
 import PostFilter from "@features/posts/ui/PostFilter.tsx";
 import PostManagementHeader from "@widgets/post/PostManagementHeader.tsx";
-import { useInitializePostStore } from "@pages/model/useInitializePost.ts";
+import { useQueryPost } from "@features/posts/model/useInitializePost.ts";
 import PostPagination from "@features/posts/ui/PostPagination.tsx";
 import LoadingIndicator from "@shared/indicator/ui/LoadingIndicator.tsx";
 
 const PostsManager = () => {
-  const { isLoading } = useInitializePostStore();
+  const { isLoading } = useQueryPost();
 
   return (
     <Card className="w-full max-w-6xl mx-auto">
